+++ BelissimoCloneWPF.Domain/Commons/Auditable.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BelissimoCloneWPF.Domain.Commons
+﻿namespace BelissimoCloneWPF.Domain.Commons
 {
     public abstract class Auditable
     {
