+++ BelissimoCloneWPF.Domain/Entities/Orders/Basket.cs
@@ -1,8 +1,8 @@
 ﻿using BelissimoCloneWPF.Domain.Commons;
 
-namespace BelissimoCloneWPF.Domain.Entities.Baskets
+namespace BelissimoCloneWPF.Domain.Entities.Orders
 {
-    public sealed class Basket: Auditable
+    public sealed class Basket : Auditable
     {
         public int UserId { get; set; }
         public int TotalPrice { get; set; }
