+++ BelissimoCloneWPF.Domain/Entities/Orders/FoodOrder.cs
@@ -1,5 +1,4 @@
 ﻿using BelissimoCloneWPF.Domain.Commons;
-using BelissimoCloneWPF.Domain.Entities.Baskets;
 
 namespace BelissimoCloneWPF.Domain.Entities.Orders
 {
