+++ BelissimoCloneWPF.Domain/Entities/Orders/Order.cs
@@ -1,5 +1,4 @@
 ﻿using BelissimoCloneWPF.Domain.Commons;
-using BelissimoCloneWPF.Domain.Entities.Baskets;
 using BelissimoCloneWPF.Domain.Entities.Branches;
 using BelissimoCloneWPF.Domain.Entities.Users;
 using BelissimoCloneWPF.Domain.Enums;
