+++ BelissimoCloneWPF.Domain/Enums/UserRole.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BelissimoCloneWPF.Domain.Enums
+﻿namespace BelissimoCloneWPF.Domain.Enums
 {
     public class UserRole
     {
